Uppgift 4: -Lagt tillen klass som tar hand om inmatningar
