+++ Uppgift4/ArvOchAbstraktion/VerkstadV2.cs
@@ -1,0 +1,25 @@
+﻿using ArvOchAbstraktion;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Klasser
+{
+    class VerkstadV2 : IVerkstad
+    {
+        public void AddVehicle(Vehicle vehicle)
+        {
+            throw new NotImplementedException();
+        }
+
+        public List<Vehicle> GetListOfVehicles()
+        {
+            throw new NotImplementedException();
+        }
+
+        public void RemoveVehicle(Vehicle vehicle)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
