+++ Uppgift4/Klasser/Car.cs
@@ -8,12 +8,6 @@ namespace Klasser
     {
         public bool HasTowbar { get; set; }
 
-        //public Car(string modelName, string licensePlate, string registrationDate, bool hasTowbar) : base(modelName, licensePlate, registrationDate)
-        //{
-        //    HasTowbar = hasTowbar;
-        //    _typeOfVehicle = "Bil";
-        //}
-
         public Car()
         {
             _typeOfVehicle = "Bil";
@@ -27,9 +21,9 @@ namespace Klasser
         {
             base.PrintInfo();
             if (HasTowbar)
-                Console.WriteLine("Dragkrok: JA");
+                Console.WriteLine("Dragkrok: Ja");
             else
-                Console.WriteLine("Dragkrok: NEJ");
+                Console.WriteLine("Dragkrok: Nej");
         }
 
     }
